% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifyWafers.R
\name{classifyWafers}
\alias{classifyWafers}
\title{Classify one or more wafers.}
\usage{
classifyWafers(distributions, trainedKohonenObject)
}
\arguments{
\item{distributions}{a matrix where any row represent the smoothed fault probability function
on a wafer.}

\item{trainedKohonenObject}{a \code{kohonen} object trained with wafers.}
}
\value{
vector where every element is the cluster of a wafer.
}
\description{
Classify one or more wafers according to a trained kohonen object.
}

