% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{distributions}
\alias{distributions}
\title{Performs KDE on a suitable data.frame}
\usage{
distributions(KDEdataFrame, features)
}
\arguments{
\item{KDEdataFrame}{a data.frame containin all the faults. The format should be the
one specified in the description.}

\item{features}{a list containing the dimensional features of the wafer. Required fields are:
dieWidth: the width of a chip
dieHeight: the height of a chip
waferRay: the ray of the wafer}
}
\value{
a matrix where any row represent the smoothed fault probability function on a wafer.
}
\description{
Load data from a data frame and performs KDE on all wafers. It it assumed that it has three column: the one
named ID contains the name of the wafer, the one named FAULT_X contains the
x coord of the fault, while the FAULT_Y column contains the y coord of the
faulty chip.
}

