% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/waferFaultMap.R
\name{waferFaultMap}
\alias{waferFaultMap}
\title{Build the fault map for a wafer.}
\usage{
waferFaultMap(features, index, faultValue = 1, nonFaultValue = 0,
  outValue = NA)
}
\arguments{
\item{features}{a list containing the dimensional features of the wafer. Required fields are:
dieWidth: the width of a chip
dieHeight: the height of a chip
waferRay: the ray of the wafer}

\item{index}{a matrix whise columns are the index (ie. 1 die counts 1) of faults.
These numbers are assumed to be index of a matrix, so they jhave to be positive integers.}

\item{faultValue}{the value to represent a faulty chip. Default is 1}

\item{nonFaultValue}{the value to represent a good chip. Default is 0}

\item{outValue}{the value to represent a chip outside the circular wafer. Default is NA}
}
\value{
a matrix whose elements represent dies on the wafer. The value of each element is a conventional
symbol for its condition (good, bad, out of wafer).
}
\description{
This function builds a fault map for a selected wafer.
}

