% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/waferKDEMap.R
\name{waferKDEMap}
\alias{waferKDEMap}
\title{KDE algorithm on a single wafer.}
\usage{
waferKDEMap(features, coords)
}
\arguments{
\item{features}{a list containing the dimensional features of the wafer. Required fields are:
dieWidth: the width of a chip
dieHeight: the height of a chip
waferRay: the ray of the wafer}

\item{coords}{a matrix whise columns are the logical (ie. 1 die counts 1) coordinates of faults}
}
\value{
a matrix whose elements represent the probability for each wafer to be faulty.
}
\description{
This function perform KDE algorithm on a single wafer.
}

