% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/classificationFrame.R
\name{classificationFrame}
\alias{classificationFrame}
\title{Writes data.frame with clustering data.}
\usage{
classificationFrame(trainedSomObject, KDEFrame)
}
\arguments{
\item{trainedSomObject}{a kohonen object trained with the wafer in the KDEFrame}

\item{KDEFrame}{a data frame with three columns: the wafer Id in the first column and the X-Y
coordiantes of the faults in the other two columns}
}
\value{
A data frame containing the name of the wafer, the number of faulty chips and the cluster.
}
\description{
This function returns a data.frame with the clustering of the data passed as an argument.
}

